+++ LotComWatcher/Program.cs
@@ -1,3 +1,5 @@
+using LotCom.Database.Auth;
+using LotCom.Database.Http;
 using LotComWatcher;
 
 var builder = Host.CreateApplicationBuilder(args);
@@ -6,6 +8,14 @@ builder.Services.AddWindowsService(options =>
 {
     options.ServiceName = "LotCom Watcher microservice";
 });
+
+// inject an HttpClient for the app
+UserAgent Agent = UserAgentFactory.CreateWatcherAgent(System.Reflection.Assembly.GetEntryAssembly()!.GetName().Version!.ToString());
+builder.Services.AddSingleton(Agent);
+HttpClient Http = HttpClientFactory.Create(Agent);
+builder.Services.AddSingleton(Http);
+
+// inject the Worker service
 builder.Services.AddHostedService<Worker>();
 
 var host = builder.Build();